components divided
