+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 
 import Footer from "./components/Footer";
-
+import Activity from "./components/Activity";
 import Profile from "./components/Profile";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap-icons/font/bootstrap-icons.css";
